Add Timer Alarm to RecipeDetailTimer
