+++ flutter/ottugi_curry/lib/view/controller/recipe_detail/recipe_detail_timer_controller.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:assets_audio_player/assets_audio_player.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 
@@ -12,6 +13,8 @@ class RecipeDetailTimerController {
   TextEditingController secondTextEditingController = TextEditingController(text: '00');
   RxString startButtonText = '시작'.obs;
 
+  final AssetsAudioPlayer timerAlarmPlayer = AssetsAudioPlayer.newPlayer();
+  RxBool isRingingAlarm = false.obs;
 
   void startTimer() { //시작, 재개
     // 입력한 시간초 기반 total Seconds 설정
@@ -30,6 +33,9 @@ class RecipeDetailTimerController {
       print('remainingSeconds: $remainingSeconds');
       if (remainingSeconds.value == 0) {
         stopTimer();
+        // 알림 시작
+        timerAlarmPlayer.play();
+        isRingingAlarm.value = true;
       } else {
         // 타이머 시간초 텍스트 변경
         minuteTextEditingController.text = (remainingSeconds.value ~/ 60).toString().padLeft(2, "0");
@@ -48,11 +54,12 @@ class RecipeDetailTimerController {
     // 타이머 종료
     isRunning.value = false;
     timer?.cancel();
+    // 알람 종료
+    stopTimerAlarm();
 
     // start 버튼 텍스트 변경
     startButtonText.value = '시작';
   }
-
 
   void pauseTimer() { // 일시정지
     // 타이머 종료 - isRunning == true
@@ -60,5 +67,21 @@ class RecipeDetailTimerController {
 
     // start 버튼 텍스트 변경
     startButtonText.value = '재개';
+  }
+
+  void loadTimerAlarm() {
+    timerAlarmPlayer.open(
+      Audio("assets/audios/timerAlarm.mp3"),
+      loopMode: LoopMode.single, //반복 여부 (LoopMode.none : 없음)
+      autoStart: false, //자동 시작 여부
+      showNotification: false, //스마트폰 알림 창에 띄울지 여부
+    );
+  }
+
+  void stopTimerAlarm() {
+    if (isRingingAlarm.value) {
+      timerAlarmPlayer.stop();
+      isRingingAlarm.value = false;
+    }
   }
 }
