+++ flutter/ottugi_curry/pubspec.yaml
@@ -45,6 +45,7 @@ dependencies:
   json_annotation: ^4.8.1
   flutter_secure_storage: ^8.0.0
   localstorage: ^4.0.1+2
+  assets_audio_player: ^3.0.6
 
 dev_dependencies:
   flutter_test:
@@ -72,6 +73,7 @@ flutter:
     - assets/
     - assets/images/
     - assets/icons/
+    - assets/audios/
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
